所得限度額の記述を修正
