+++ openfisca_japan/parameters/福祉/育児/児童手当/所得制限限度額.yaml
@@ -1,42 +1,14 @@
-description: 渋谷区の児童手当の所得制限限度額
+description: 渋谷区の児童手当の所得制限限度額(扶養人数が0人から増えるごとに増額)
 metadata:
   threshold_unit: person
   rate_unit: 1
   reference: https://www.city.shibuya.tokyo.jp/kodomo/teate/teate/teate_gendogaku.html
-brackets:
-  - rate:
-      2000-05-01:
-        value: 6220000
-    threshold:
-      2000-05-01:
-        value: 0
-  - rate:
-      2000-05-01:
-        value: 6600000
-    threshold:
-      2000-05-01:
-        value: 1
-  - rate:
-      2000-05-01:
-        value: 6980000
-    threshold:
-      2000-05-01:
-        value: 2
-  - rate:
-      2000-05-01:
-        value: 7360000
-    threshold:
-      2000-05-01:
-        value: 3
-  - rate:
-      2000-05-01:
-        value: 7740000
-    threshold:
-      2000-05-01:
-        value: 4
-  - rate:
-      2000-05-01:
-        value: 8120000
-    threshold:
-      2000-05-01:
-        value: 5
+values:
+  2000-05-01:
+    value:
+      - 6220000
+      - 6600000
+      - 6980000
+      - 7360000
+      - 7740000
+      - 8120000
