+++ openfisca_japan/tests/年齢と学年/学年.yaml
@@ -3,9 +3,6 @@
   input:
     誕生年月日: 2010-04-02
   output:
-    年齢:
-      2018-04: 7
-      2018-05: 8
     学年:
       2018-03-31: 1
       2018-04-01: 2
@@ -15,9 +12,6 @@
   input:
     誕生年月日: 2010-04-01
   output:
-    年齢:
-      2018-04: 8
-      2018-05: 8
     学年:
       2017-03-31: 1
       2017-04-01: 2
