+++ openfisca_japan/tests/年齢と学年/年齢.yaml
@@ -1,13 +1,12 @@
-
 - name: 生まれた年の年齢は0歳である
   period: 2016-01
   input:
     誕生年月日: 1980-02-01
   output:
     年齢:
-      1980-02: 0
-      1980-03: 0
-      1981-03: 1
+      1980-02-01: 0
+      1980-03-01: 0
+      1981-02-01: 1
 
 - name: 誕生日を過ぎたら年齢が上がる
   period: 2022-06
@@ -15,8 +14,8 @@
     誕生年月日: 1980-02-15
   output:
     年齢:
-      2015-02: 34
-      2015-03: 35
+      2015-02-14: 34
+      2015-02-15: 35
 
 - name: 大晦日に生まれた人も年齢が上がる
   period: 2016-01
@@ -24,8 +23,8 @@
     誕生年月日: 1980-12-31
   output:
     年齢:
-      2015-12: 34
-      2016-01: 35
+      2015-12-31: 35
+      2016-01-01: 35
 
 - name: 誕生日が閏日でも適切に年齢が上がる
   period: 2022-01
@@ -33,5 +32,5 @@
     誕生年月日: 2000-02-29
   output:
     年齢:
-      2015-02: 14
-      2015-03: 15
+      2015-02-28: 14
+      2015-03-01: 15
