+++ openfisca_japan/variables/人口.py
@@ -40,7 +40,8 @@ class 死亡年月日(Variable):
 class 年齢(Variable):
     value_type = int
     entity = 人物
-    definition_period = MONTH
+    #definition_period = MONTH
+    definition_period = DAY
     label = "人物の年齢"
 
     def formula(対象人物, 対象期間, _parameters):
