+++ openfisca_japan/variables/福祉.py
@@ -7,7 +7,7 @@ See https://openfisca.org/doc/key-concepts/variables.html
 """
 
 # Import from openfisca-core the Python objects used to code the legislation in OpenFisca
-from openfisca_core.periods import MONTH
+from openfisca_core.periods import MONTH, DAY
 from openfisca_core.variables import Variable
 # Import the Entities specifically defined for this tax and benefit system
 from openfisca_japan.entities import 世帯, 人物
@@ -58,25 +58,3 @@ class 年金(Variable):
     def formula(対象人物, 対象期間, parameters):
         年齢条件 = 対象人物("年齢", 対象期間) >= parameters(対象期間).全般.定年年齢
         return 年齢条件
-
-
-class 世帯所得(Variable):
-    value_type = float
-    entity = 世帯
-    definition_period = MONTH
-    label = "世帯全員の収入の合計"
-
-    def formula(対象世帯, 対象期間, _parameters):
-        各収入 = 対象世帯.members("所得", 対象期間)
-        return 対象世帯.sum(各収入)
-
-
-class 世帯高所得(Variable):
-    value_type = float
-    entity = 世帯
-    definition_period = MONTH
-    label = "世帯で最も所得が高い人物の所得"
-
-    def formula(対象世帯, 対象期間, _parameters):
-        各収入 = 対象世帯.members("所得", 対象期間)
-        return 対象世帯.max(各収入)
