+++ openfisca_japan/variables/福祉/育児/児童手当.py
@@ -35,28 +35,23 @@ class 児童手当(Variable):
         # 扶養親族に配偶者は含まれない。(親等の児童以外を扶養する場合はそれらも含む必要あり)
         # 扶養親族の定義(参考): https://www.nta.go.jp/taxes/shiraberu/taxanswer/shotoku/1180.htm
         扶養親族である = 児童である * (世帯所得一覧 < 扶養控除所得金額)
-        扶養人数 = 対象世帯.sum(扶養親族である)
-        所得制限限度額 = 児童手当.所得制限限度額.calc(扶養人数)
-        所得上限限度額 = 児童手当.所得上限限度額.calc(扶養人数)
+        扶養人数 = int(対象世帯.sum(扶養親族である)[0])
+        所得制限限度額 = 児童手当.所得制限限度額[扶養人数]
+        所得上限限度額 = 児童手当.所得上限限度額[扶養人数]
         児童手当所得条件 = 世帯高所得 < 所得制限限度額
         特例給付所得条件 = (世帯高所得 >= 所得制限限度額) * (世帯高所得 < 所得上限限度額)
-
-        print(世帯所得一覧)
-        print(f'扶養人数: {扶養人数}')
-        print(対象世帯.has_role(世帯.児童))
 
         年齢 = 対象世帯.members("年齢", 対象期間)
         学年 = 対象世帯.members("学年", 対象期間)
         三歳未満の人数 = 対象世帯.sum(年齢 < 3)
         学年の降順インデックス = np.argsort(学年)[::-1]
-        高校生修了後の人数 = 対象世帯.sum(学年 > 12)
+        高校生修了後の人数 = int(対象世帯.sum(学年 > 12)[0])
         高校生以下で数えた第三子以降である = np.full(len(年齢), False)
         if 高校生修了後の人数+2 < len(年齢):
             高校生以下で数えた第三子以降である[学年の降順インデックス[高校生修了後の人数+2:]] = True
 
         三歳から小学校修了前かつ第二子以前の人数 = 対象世帯.sum((年齢 >= 3) * (学年 <= 6) * np.logical_not(高校生以下で数えた第三子以降である))
         三歳から小学校修了前かつ第三子以降の人数 = 対象世帯.sum((年齢 >= 3) * (学年 <= 6) * 高校生以下で数えた第三子以降である)
-        
         中学生の人数 = 対象世帯.sum((学年 >= 7) * (学年 <= 9))
 
         児童手当金額 = \
